Renaming base imaging script for consistency with other base classes
