+++ src/synthesizer/imaging/base_imaging.py
