+++ src/synthesizer/imaging/image.py
@@ -24,12 +24,12 @@ from scipy.ndimage import zoom
 from unyt import arcsecond, kpc, unyt_array, unyt_quantity
 
 from synthesizer import exceptions
+from synthesizer.imaging.base_imaging import ImagingBase
 from synthesizer.imaging.image_generators import (
     _generate_image_parametric_smoothed,
     _generate_image_particle_hist,
     _generate_image_particle_smoothed,
 )
-from synthesizer.imaging.imaging_base import ImagingBase
 from synthesizer.units import accepts, unit_is_compatible
 from synthesizer.utils import TableFormatter
